Revert "Enhance/meta 4121"
