+++ intg/src/main/java/org/apache/atlas/model/lineage/AtlasLineageInfo.java
@@ -19,7 +19,6 @@ package org.apache.atlas.model.lineage;
 
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -279,8 +278,6 @@ public class AtlasLineageInfo implements Serializable {
         @JsonProperty
         boolean                    hasDownstream;
         LineageOnDemandConstraints onDemandConstraints;
-        @JsonIgnore
-        int                        offsetCount;
 
         public LineageInfoOnDemand() { }
 
@@ -294,15 +291,6 @@ public class AtlasLineageInfo implements Serializable {
             this.isOutputRelationsReachedLimit = false;
             this.hasUpstream                   = false;
             this.hasDownstream                 = false;
-            this.offsetCount                   = 0;
-        }
-
-        public int getOffsetCount() {
-            return offsetCount;
-        }
-
-        public int incrementOffsetCount() {
-            return offsetCount++;
         }
 
         public boolean isInputRelationsReachedLimit() {
